requirements.txt
fastapi
uvicorn
openai
sqlalchemy
python-dotenv
faiss-cpu
sentence-transformers
numpy